Add support for caching of the parse_sub() results.
This is the first stage of caching, there could be one or two
more (generated sql caching and prepared statements caching).

It is incomplete and not usable in production - the existing tests do
not all pass, and there are no tests to check how well the caching
works.

Known problems: comparison tests fail, some overloading problems
with the new Opval class.

Expected to be a problem:
- subselects, explicit joins, unions & friends will all get the old padlist;
- variable table names are not taken into account.